requirements.txt
fastapi
uvicorn
motor
python-jose
bcrypt
pydantic
